added edit command
Enable editing a chosen line in an existing program:
edit  <n>
where n is a valid line number
note that it doesn't remove the original line if you change the line number in edit buffer.

also fixed oops in int_to_sz which never print a single '0' because of leading zero filter.